requirements.txt
fastapi~=0.115.12
pydantic~=2.11.5
mysql-connector-python~=9.3.0
numpy~=2.2.4
opencv-python~=4.11.0.86
pandas~=2.2.3
uvicorn~=0.34.3
python-multipart~=0.0.5

scikit-image~=0.25.2